Move color functions to own compile unit
