+++ Makefile.conf
@@ -11,6 +11,7 @@ SOURCES = \
 	context.cpp \
 	constants.cpp \
 	fn_utils.cpp \
+	fn_colors.cpp \
 	fn_numbers.cpp \
 	fn_selectors.cpp \
 	functions.cpp \
