+++ win/libsass.vcxproj.filters
@@ -114,6 +114,9 @@
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\functions.hpp">
       <Filter>Headers</Filter>
     </ClInclude>
+    <ClInclude Include="$(LIBSASS_HEADERS_DIR)\fn_colors.hpp">
+      <Filter>Headers</Filter>
+    </ClInclude>
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\fn_numbers.hpp">
       <Filter>Headers</Filter>
     </ClInclude>
@@ -279,6 +282,9 @@
       <Filter>Sources</Filter>
     </ClCompile>
     <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_utils.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_colors.cpp">
       <Filter>Sources</Filter>
     </ClCompile>
     <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_numbers.cpp">
